Aggiunto JAAS LoginModule custom,JAASLoginInterceptor, WSS4JInInterceptor, SecureAnnotationsInterceptor. Modifica Service.

+++ src/main/java/it/univaq/sose/simplebankingsoapservice/webservice/BankWebServiceImpl.java
@@ -8,9 +8,14 @@ import it.univaq.sose.simplebankingsoapservice.dto.MoneyTransfer;
 import it.univaq.sose.simplebankingsoapservice.dto.OpenBankAccountRequest;
 import it.univaq.sose.simplebankingsoapservice.repository.AccountRepository;
 import it.univaq.sose.simplebankingsoapservice.repository.BankAccountRepository;
+import it.univaq.sose.simplebankingsoapservice.security.AccountDetails;
+import it.univaq.sose.simplebankingsoapservice.util.AuthenticationUtils;
+import org.apache.cxf.interceptor.security.AuthenticationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.annotation.Resource;
+import javax.xml.ws.WebServiceContext;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
@@ -20,9 +25,16 @@ public class BankWebServiceImpl implements BankWebService {
     private BankAccountRepository bankAccountRepository = BankAccountRepository.getInstance();
     private static final Logger LOG = LoggerFactory.getLogger(BankWebServiceImpl.class);
 
+    @Resource
+    private WebServiceContext wsContext;
+
     @Override
     public AccountAndBankAccount getAccountAndBankAccount(long accountId) throws NotFoundException {
+        AccountDetails authenticationDetails = AuthenticationUtils.getAuthenticationDetails(wsContext);
         Account account = accountRepository.findById(accountId);
+        if (authenticationDetails.getRole().equals(Role.CUSTOMER) && !account.getUsername().equals(authenticationDetails.getUsername())) {
+            throw new AuthenticationException("You are not authorised to make this request");
+        }
         BankAccount bankAccount = bankAccountRepository.findById(account.getIdBankAccount());
         LOG.info("Risposta getAccountAndBankAccount");
         AccountAndBankAccount accountAndBankAccount = new AccountAndBankAccount(account.getIdAccount(), account.getName(), account.getSurname(), account.getUsername(), bankAccount);
@@ -79,9 +91,14 @@ public class BankWebServiceImpl implements BankWebService {
 
     @Override
     public AccountAndBankAccount depositMoneyInBankAccount(MoneyTransfer moneyTransfer) throws NotFoundException {
-        bankAccountRepository.addMoney(moneyTransfer.getIdBankAccount(), moneyTransfer.getAmount());
+        AccountDetails authenticationDetails = AuthenticationUtils.getAuthenticationDetails(wsContext);
         BankAccount bankAccount = bankAccountRepository.findById(moneyTransfer.getIdBankAccount());
         Account account = accountRepository.findById(bankAccount.getAccount());
+        if (authenticationDetails.getRole().equals(Role.CUSTOMER) && !account.getUsername().equals(authenticationDetails.getUsername())) {
+            throw new AuthenticationException("You are not authorised to make this request");
+        }
+        bankAccountRepository.addMoney(moneyTransfer.getIdBankAccount(), moneyTransfer.getAmount());
+        bankAccount = bankAccountRepository.findById(moneyTransfer.getIdBankAccount());
         AccountAndBankAccount accountAndBankAccount = new AccountAndBankAccount(account.getIdAccount(), account.getName(), account.getSurname(), account.getUsername(), bankAccount);
         LOG.info("Risposta depositMoneyInBankAccount");
         LOG.info("{}", accountAndBankAccount);
@@ -90,9 +107,14 @@ public class BankWebServiceImpl implements BankWebService {
 
     @Override
     public AccountAndBankAccount withdrawMoneyInBankAccount(MoneyTransfer moneyTransfer) throws NotFoundException, InsufficientFundsException {
-        bankAccountRepository.removeMoney(moneyTransfer.getIdBankAccount(), moneyTransfer.getAmount());
+        AccountDetails authenticationDetails = AuthenticationUtils.getAuthenticationDetails(wsContext);
         BankAccount bankAccount = bankAccountRepository.findById(moneyTransfer.getIdBankAccount());
         Account account = accountRepository.findById(bankAccount.getAccount());
+        if (authenticationDetails.getRole().equals(Role.CUSTOMER) && !account.getUsername().equals(authenticationDetails.getUsername())) {
+            throw new AuthenticationException("You are not authorised to make this request");
+        }
+        bankAccountRepository.removeMoney(moneyTransfer.getIdBankAccount(), moneyTransfer.getAmount());
+        bankAccount = bankAccountRepository.findById(moneyTransfer.getIdBankAccount());
         AccountAndBankAccount accountAndBankAccount = new AccountAndBankAccount(account.getIdAccount(), account.getName(), account.getSurname(), account.getUsername(), bankAccount);
         LOG.info("Risposta withdrawMoneyInBankAccount");
         LOG.info("{}", accountAndBankAccount);
